Move `m_pLoadBestTimeResult` to `IGameController`
Gets rid of more CGameControllerDDRace casts.
Those did not show up during compile time only during runtime
of another active controller when asan is active.
