+++ src/game/server/gamemodes/DDRace.h
@@ -4,7 +4,6 @@
 
 #include <game/server/gamecontroller.h>
 
-struct CScoreLoadBestTimeResult;
 class CGameControllerDDRace : public IGameController
 {
 public:
@@ -23,7 +22,5 @@ public:
 	void Tick() override;
 
 	void DoTeamChange(class CPlayer *pPlayer, int Team, bool DoChatMsg = true) override;
-
-	std::shared_ptr<CScoreLoadBestTimeResult> m_pLoadBestTimeResult;
 };
 #endif // GAME_SERVER_GAMEMODES_DDRACE_H
